requirements.txt
# Sanchalak - Government Scheme Eligibility System
# Core dependencies for the complete pipeline

# Web Framework and API
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Frontend
streamlit>=1.46.0
streamlit-webrtc>=0.47.0

# Database
pymongo>=4.0.0
redis>=5.0.0

# HTTP and Networking
requests>=2.32.0
httpx>=0.25.0

# Data Processing
numpy>=2.3.0
pandas>=2.3.0
pyarrow>=20.0.0

# Audio Processing
pydub>=0.25.0
av>=10.0.0

# Configuration and Environment
python-dotenv>=1.0.0
toml>=0.10.0

# Logging and Monitoring
structlog>=23.0.0
prometheus-client>=0.19.0

# YAML and JSON Processing
pyyaml>=6.0
jsonschema>=4.24.0

# Machine Learning and AI
torch>=2.0.0
transformers>=4.35.0
experta>=1.9.0

# Prolog Integration (optional)
pyswip>=0.2.10

# System Monitoring
psutil>=5.9.0
GPUtil>=1.4.0

# Web Components
jinja2>=3.1.0
tornado>=6.5.0

# Streamlit Components
altair>=5.5.0
pydeck>=0.9.0

# Utilities
click>=8.2.0
blinker>=1.9.0
tenacity>=9.1.0
typing-extensions>=4.14.0
protobuf>=5.29.0

# Git Integration
gitpython>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Async Support
anyio>=4.9.0

# Web Scraping (if needed)
beautifulsoup4>=4.13.0

# Additional Dependencies
pillow>=11.2.0 